[IMP]Split js in two files and updated module name as namespace.
