+++ web_group_expand/templates/assets.xml
@@ -4,6 +4,7 @@
         <xpath expr="." position="inside">
             <link rel="stylesheet" href="/web_group_expand/static/src/css/web_group_expand.css"/>
             <script type="text/javascript" src="/web_group_expand/static/src/js/web_group_expand.js"></script>
+            <script type="text/javascript" src="/web_group_expand/static/src/js/web_group_expand_menu.js"></script>
         </xpath>
     </template>
 </odoo>
